feat(daemon): use dialoguer to handle prompts
